+++ tools/utils.c
@@ -1,8 +1,8 @@
 #include "tools/utils.h"
 
-bool ptrInArray(void* ptr, void* _array[], unsigned int size)
+bool ptrInArray(void* ptr, void* _array[], size_t size)
 {
-	for (unsigned int i = 0; i < size; i++)
+	for (uint32_t i = 0; i < size; i++)
 	{
 		if (_array[i] == ptr)
 			return true;
